+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## 2.1.0 - Not yet released
 - Adds paragraph line feed to textual ViewHelpers, [#107](https://github.com/bithost-gmbh/pdfviewhelpers/issues/107) (Thanks [@emmemme](https://github.com/emmemme))
-- Adds image processing to ImageViewHelper
+- Adds image processing to ImageViewHelper, [#104](https://github.com/bithost-gmbh/pdfviewhelpers/issues/104)
+- Adds table of content ViewHelpers, [#103](https://github.com/bithost-gmbh/pdfviewhelpers/issues/103)
 - Updates documentation
 
 ## 2.0.2 - December 1, 2018